reorganize to use as lib
