+++ .travis.yml
@@ -14,7 +14,7 @@ script:
   # Creating default config
   - ./clickhouse-backup/clickhouse-backup default-config > clickhouse-backup/config.yml || travis_terminate 1;
   # Running unit tests
-  - go test -v || travis_terminate 1;
+  - go test -v ./pkg/...  || travis_terminate 1;
   # Decrypting credentials for Google Cloud Storage
   - openssl enc -base64 -aes-256-cbc -d -in integration-test/credentials.json.enc -out integration-test/credentials.json -k ${VAULT_PASSWORD} || travis_terminate 1;
   # Preparing to integration tests
