+++ go.mod
@@ -1,30 +1,40 @@
 module github.com/AlexAkulov/clickhouse-backup
 
 require (
-	cloud.google.com/go/storage v1.1.0
-	github.com/aws/aws-sdk-go v1.23.21
-	github.com/bkaradzic/go-lz4 v1.0.0 // indirect
-	github.com/cloudflare/golz4 v0.0.0-20150217214814-ef862a3cdc58 // indirect
+	cloud.google.com/go v0.47.0 // indirect
+	cloud.google.com/go/storage v1.1.2
+	github.com/andybalholm/brotli v0.0.0-20190821151343-b60f0d972eeb // indirect
+	github.com/aws/aws-sdk-go v1.25.19
+	github.com/cpuguy83/go-md2man/v2 v2.0.0 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/djherbis/buffer v1.1.0 // indirect
 	github.com/fatih/color v1.7.0 // indirect
 	github.com/frankban/quicktest v1.5.0 // indirect
+	github.com/golang/groupcache v0.0.0-20191027212112-611e8accdfc9 // indirect
 	github.com/jmoiron/sqlx v1.2.0
-	github.com/kelseyhightower/envconfig v1.3.0
-	github.com/kshvakov/clickhouse v1.3.7
+	github.com/jstemmer/go-junit-report v0.9.1 // indirect
+	github.com/kelseyhightower/envconfig v1.4.0
+	github.com/klauspost/compress v1.9.1 // indirect
+	github.com/kshvakov/clickhouse v1.3.11
 	github.com/mattn/go-colorable v0.1.1 // indirect
-	github.com/mattn/go-runewidth v0.0.3 // indirect
+	github.com/mattn/go-runewidth v0.0.5 // indirect
 	github.com/mholt/archiver v1.1.3-0.20190812163345-2d1449806793
 	github.com/pierrec/lz4 v2.3.1-0.20190907084430-0554902aa74a+incompatible // indirect
 	github.com/pkg/errors v0.8.1
 	github.com/stretchr/testify v1.3.0
-	github.com/urfave/cli v1.20.0
-	golang.org/x/tools v0.0.0-20190918214516-5a1a30219888 // indirect
-	google.golang.org/api v0.9.0
+	github.com/urfave/cli v1.22.1
+	go.opencensus.io v0.22.1 // indirect
+	golang.org/x/exp v0.0.0-20191024150812-c286b889502e // indirect
+	golang.org/x/net v0.0.0-20191028085509-fe3aa8a45271 // indirect
+	golang.org/x/sys v0.0.0-20191027211539-f8518d3b3627 // indirect
+	golang.org/x/tools v0.0.0-20191026034945-b2104f82a97d // indirect
+	google.golang.org/api v0.11.0
+	google.golang.org/appengine v1.6.5 // indirect
+	google.golang.org/grpc v1.24.0 // indirect
 	gopkg.in/cheggaaa/pb.v1 v1.0.28
 	gopkg.in/djherbis/buffer.v1 v1.1.0
 	gopkg.in/djherbis/nio.v2 v2.0.3
-	gopkg.in/yaml.v2 v2.2.1
+	gopkg.in/yaml.v2 v2.2.4
 )
 
 go 1.13
