+++ pkg/chbackup/s3.go
@@ -1,4 +1,4 @@
-package main
+package chbackup
 
 import (
 	"io"
