+++ pkg/chbackup/utils.go
@@ -1,4 +1,4 @@
-package main
+package chbackup
 
 import (
 	"fmt"
