+++ pkg/chbackup/utils_test.go
@@ -1,4 +1,4 @@
-package main
+package chbackup
 
 import (
 	"log"
